requirements.txt
requests
keyring
python-dateutil